# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Web Manga API
  version: 0.2.0
  description: API served by manga-service from PostgreSQL gold tables.
servers:
  - url: http://localhost:8000
paths:
  /health:
    get:
      summary: Service health check
      responses:
        "200":
          description: Service status
  /manga:
    get:
      summary: List manga from gold catalog
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: original_language
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated manga catalog items
  /manga/{manga_id}:
    get:
      summary: Get manga detail from gold detail table
      parameters:
        - name: manga_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Manga detail
        "404":
          description: Manga not found
  /manga/{manga_id}/chapters:
    get:
      summary: List chapters for a manga
      parameters:
        - name: manga_id
          in: path
          required: true
          schema:
            type: string
        - name: translated_language
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated chapter list
  /chapters/latest:
    get:
      summary: List latest chapters
      parameters:
        - name: translated_language
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated latest chapters